refactor(components/input): 重构 Input 组件，抽离出 InputSkin 组件
