+++ packages/components/src/input/components/index.ts
@@ -1,2 +1,3 @@
-export * from './InnerInput';
-export * from './Suffix';
+export * from './input.Switch';
+export * from './input.Suffix';
+export * from './input.Skin';
