+++ packages/components/src/input/index.scss
@@ -1,13 +1,14 @@
 @use '../namespace' as Name;
 
 $r: Name.$input;
+$skin: Name.$input-skin;
 
-.#{$r} {
-  --t-select-border-color: var(--t-color-info-a-3);
-  --t-select-box-shadow: 0 0 0 0 var(--t-color-primary-a-6);
-  --t-select-bg-color: initial;
-  --t-select-active-border-color: var(--t-color-primary-a-2);
-  --t-select-active-box-shadow: 0 0 0 2px var(--t-color-primary-a-6);
+.#{$skin} {
+  --t-input-border-color: var(--t-color-info-a-3);
+  --t-input-box-shadow: 0 0 0 0 var(--t-color-primary-a-6);
+  --t-input-bg-color: initial;
+  --t-input-active-border-color: var(--t-color-primary-a-2);
+  --t-input-active-box-shadow: 0 0 0 2px var(--t-color-primary-a-6);
 
   position: relative;
   display: flex;
@@ -15,16 +16,48 @@ $r: Name.$input;
   padding: 2px var(--t-radius);
   min-width: 120px;
   min-height: var(--t-size);
-  border: 1px solid var(--t-select-border-color);
+  border: 1px solid var(--t-input-border-color);
   border-radius: var(--t-radius);
-  background-color: var(--t-select-bg-color);
-  box-shadow: var(--t-select-box-shadow);
+  background-color: var(--t-input-bg-color);
+  box-shadow: var(--t-input-box-shadow);
   gap: 2px;
   transition-property: color, background-color, border-color, box-shadow;
   transition-duration: var(--t-transition-duration);
   box-sizing: border-box;
   line-height: 1;
   cursor: text;
+
+  // --
+  &--active,
+  &--focus {
+    border-color: var(--t-input-active-border-color);
+    box-shadow: var(--t-input-active-box-shadow);
+  }
+  &--warning {
+    --t-input-border-color: var(--t-color-warning-a-2);
+    --t-input-active-border-color: var(--t-color-warning-a-2);
+    --t-input-active-box-shadow: 0 0 0 2px var(--t-color-warning-a-6);
+  }
+  &--error {
+    --t-input-border-color: var(--t-color-danger-a-2);
+    --t-input-active-border-color: var(--t-color-danger-a-2);
+    --t-input-active-box-shadow: 0 0 0 2px var(--t-color-danger-a-6);
+  }
+  &--disabled {
+    --t-input-bg-color: var(--t-color-info-a-7);
+    --t-input-border-color: var(--t-color-info-a-3);
+
+    cursor: not-allowed;
+  }
+
+  // :
+  &:not(.#{$skin}--disabled) {
+    &:hover {
+      border-color: var(--t-input-active-border-color);
+    }
+  }
+}
+.#{$r} {
   > input,
   > textarea {
     flex: 1;
@@ -48,18 +81,11 @@ $r: Name.$input;
     line-height: 1.5;
   }
 
-  // :
-  &:hover {
-    &:not(.#{$r}--disabled) {
-      border-color: var(--t-select-active-border-color);
-    }
-  }
-
   // status
   &--textarea {
     overflow: auto;
     padding: 0;
-    &:not(.#{$r}--disabled):not(.#{$r}--autosize) {
+    &:not(.#{$skin}--disabled):not(.#{$r}--autosize) {
       resize: vertical;
     }
     .#{$r}__suffix {
@@ -70,27 +96,6 @@ $r: Name.$input;
       right: 8px;
       bottom: 6px;
     }
-  }
-  &--active,
-  &--focus {
-    border-color: var(--t-select-active-border-color);
-    box-shadow: var(--t-select-active-box-shadow);
-  }
-  &--warning {
-    --t-select-border-color: var(--t-color-warning-a-2);
-    --t-select-active-border-color: var(--t-color-warning-a-2);
-    --t-select-active-box-shadow: 0 0 0 2px var(--t-color-warning-a-6);
-  }
-  &--error {
-    --t-select-border-color: var(--t-color-danger-a-2);
-    --t-select-active-border-color: var(--t-color-danger-a-2);
-    --t-select-active-box-shadow: 0 0 0 2px var(--t-color-danger-a-6);
-  }
-  &--disabled {
-    --t-select-bg-color: var(--t-color-info-a-7);
-    --t-select-border-color: var(--t-color-info-a-3);
-
-    cursor: not-allowed;
   }
   &--clearable:hover {
     .#{$r}__clear {
