+++ packages/components/src/namespace.scss
@@ -39,3 +39,4 @@ $slider: '#{Var.$prefix}slider';
 $empty: '#{Var.$prefix}empty';
 $timeline: '#{Var.$prefix}timeline';
 $calendar: '#{Var.$prefix}calendar';
+$input-skin: '#{Var.$prefix}input-skin';
